itarators for Clause and ClauseStorage

+++ src/sat.cpp
@@ -14,7 +14,7 @@ std::ostream& operator<<(std::ostream& stream, const ClauseSet& cs)
 	// binary clauses
 	for(uint32_t l = 0; l < 2*cs.varCount(); ++l)
 		for(auto b : cs.bins[l])
-			if(l <= b.toInt())
+			if(l <= b)
 				stream << Lit(l) << " " << b << " 0\n";
 
 	// long clauses
